[FIX] MSVC complains about possible loss of information between __in64_t to int conversion due to the use of Eigen 3.
git-svn-id: https://b-tk.googlecode.com/svn/BTK/trunk@1613 93be6e1e-6fcc-11de-baec-69240fdc299a

+++ Code/Common/btkMeasure.h
@@ -62,7 +62,7 @@ namespace btk
     Values& GetValues() {return this->m_Values;};
     const Values& GetValues() const {return this->m_Values;};
     void SetValues(const Values& v);
-    int GetFrameNumber() const {return this->m_Values.rows();};
+    int GetFrameNumber() const {return static_cast<int>(this->m_Values.rows());};
     void SetFrameNumber(int frameNumber);
     Pointer Clone() const {return Pointer(new Measure<d>(*this));};
